// src/types/how-it-works.ts
export type SanzaProgramme = {
  name: string;
  description: string;
};

export type HowItWorksCopy = {
  title: string;
  lead: string;
  pemfTitle: string;
  pemfBody1: string;
  pemfBody2: string;
  pemfBody3: string;
  bioTitle: string;
  bioBody1: string;
  bioBody2: string;
  programmesSectionTitle: string;
  programmesIntro: string;
  programmeColIndex: string;
  programmeColName: string;
  programmeColDesc: string;
  programmes: SanzaProgramme[];
  lightTitle: string;
  lightBody1: string;
  lightBody2: string;
  modularTitle: string;
  modularLead: string;
  modularSpecsNote: string;
  modularGeneratorTitle: string;
  modularGeneratorBullets: string[];
  modularMatTitle: string;
  modularMatBullets: string[];
  modularPadTitle: string;
  modularPadBullets: string[];
  modularTwincTitle: string;
  modularTwincBullets: string[];
  modularPenTitle: string;
  modularPenBullets: string[];
  modularHandTitle: string;
  modularHandBullets: string[];
  modularGeneratorImageAlt: string;
  modularMatImageAlt: string;
  modularPadImageAlt: string;
  modularTwincImageAlt: string;
  modularPenImageAlt: string;
  modularHandImageAlt: string;
  ctaTitle: string;
  ctaBody: string;
  ctaExploreSectors: string;
  ctaBookDemo: string;
};